% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box}
\alias{box}
\title{Box like component}
\usage{
box(..., title = NULL, collapsed = TRUE, color = "")
}
\arguments{
\item{...}{html list contents for the panel}

\item{title}{title for the box}

\item{collapsed}{defines initial state}

\item{color}{color name as defined in custom css}
}
\value{
None
}
\description{
Box like component
}
\examples{
box()

}
