% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{dsAppPanels}
\alias{dsAppPanels}
\title{Shiny panels layout}
\usage{
dsAppPanels(..., styles = "", header = NULL, title = NULL,
  debug = FALSE)
}
\arguments{
\item{...}{html list contents for the panel}

\item{styles}{custom css styles}

\item{header}{html list with custom header}

\item{title}{Title html page attribute}

\item{debug}{Used for debugging layout html}
}
\value{
None
}
\description{
Shiny panels layout
}
\examples{
dsAppPanels()

}
