% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infomessage.R
\name{infomessage}
\alias{infomessage}
\title{Infomessage to show messages}
\usage{
infomessage(..., type = "warning", id = NULL, class = NULL)
}
\arguments{
\item{...}{html list contents for the panel}

\item{type}{one of: warning, success, error or info}

\item{id}{id of html element}

\item{class}{class of html element}
}
\value{
None
}
\description{
Infomessage to show messages
}
\examples{
infomessage()

}
