% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoTooltip.R
\name{infoTooltip}
\alias{infoTooltip}
\title{Text with tooltip}
\usage{
infoTooltip(title, info, icon = "info-circle", style = "")
}
\arguments{
\item{title}{Text}

\item{info}{Message which appears when a cursor is positioned over an icon}

\item{icon}{HTML name of icon}

\item{style}{CSS styles for the container div}
}
\value{
None
}
\description{
Text with tooltip
}
\examples{
infoTooltip("Information", info = "This a content of tooltip", icon = "cloud")

}
