% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R, R/panels.R
\name{modal}
\alias{modal}
\title{Modal window}
\usage{
modal(..., title = NULL, id = NULL)

modal(..., title = NULL, id = NULL)
}
\arguments{
\item{...}{html list contents for the panel}

\item{title}{title for the modal}

\item{id}{html attribute id for the component}
}
\value{
None

None
}
\description{
Modal window

Modal window
}
\examples{
modal()
modal()
}
