% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R, R/panels.R
\name{modalButton}
\alias{modalButton}
\title{Modal button to trigger a modal}
\usage{
modalButton(modal_id = NULL, label = NULL, id = NULL)

modalButton(modal_id = NULL, label = NULL, id = NULL)
}
\arguments{
\item{modal_id}{modal id to be triggered by the button}

\item{label}{label for the button}

\item{id}{html attribute id for the button}
}
\value{
None

None
}
\description{
Modal button to trigger a modal

Modal button to trigger a modal
}
\examples{
modal()
modal()
}
