% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{topbar}
\alias{topbar}
\title{Panel component for shiny panels layout}
\usage{
topbar(..., title = NULL, image = NULL, background_color = NULL)
}
\arguments{
\item{...}{html list contents for the panel}

\item{title}{Top bar panel}

\item{image}{Logo image}

\item{background_color}{background color for top bar}
}
\value{
None
}
\description{
Panel component for shiny panels layout
}
\examples{
dsAppPanels()

}
