% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{vizControls}
\alias{vizControls}
\title{#' @export
dataControls <- function(..., label = NULL, on = NULL){
  x <- list(...)
  tagList(
    shiny::tags$form(class="settings data__settings",
                     shiny::tags$fieldset(
                       shiny::tags$legend(label),
                       x
                     )
    )
  )
}}
\usage{
vizControls(..., basic = NULL, advanced = NULL, label = NULL,
  on = NULL)
}
\description{
#' @export
dataPreview <- function(..., on = NULL){
  x <- list(...)
  tagList(
    x
  )
}
}
