% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{panel}
\alias{panel}
\title{Panel component for shiny panels layout}
\usage{
panel(
  head = NULL,
  body = NULL,
  show_footer = TRUE,
  footer = NULL,
  title = NULL,
  color = "malibu",
  id = NULL,
  collapsed = FALSE,
  width = NULL,
  ...
)
}
\arguments{
\item{head}{html for the panel header}

\item{body}{html tag list for panel body contents}

\item{show_footer}{include footer}

\item{footer}{footer contents}

\item{title}{panel title}

\item{color}{color name as defined in custom css}

\item{id}{panel div id}

\item{collapsed}{panel starts as collapsed}

\item{width}{panel width in pixels}

\item{...}{html list contents for the panel}
}
\value{
None
}
\description{
Panel component for shiny panels layout
}
\examples{
panel(title = "My title")

}
